Spike out checking location service state in PermissionsProvider
